test/system: Adjust name of tests & add some minor changes
The name of every tests was prepended with "Try to". That is too verbose
and makes the logs more busy than they already are. Instead, use this
convention only for test cases that expect the tested command to fail.

I noticed a few things in the tests and adjusted those (e.g., no asserts
on variables expecting to equal zero).